Safari-609.1.11.1/LayoutTests/http/tests/xmlhttprequest/access-control-preflight-async-method-denied-expected.txt

CONSOLE MESSAGE: Method DELETE is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?test=access-control-preflight-async-method-denied&state=method due to access control checks.
PASS: Request successfully blocked.
